separates R and python functions to make the R dependencies optional and simplify package installation and continuous integreation
