+++ src/estimators_python.py
@@ -1,14 +1,10 @@
 """
-the objective of this script is to implement estimators for mediation in
-causal inference, simulate data, and evaluate and compare estimators
+the objective of this module is to implement estimators for mediation in
+causal inference, purely in python, without R dependency
 """
 
-# first step, run r code to have the original implementation by Huber
-# using rpy2 to have the same data in R and python...
 
-import rpy2.robjects as robjects
-import rpy2.robjects.packages as rpackages
-from rpy2.robjects import pandas2ri, numpy2ri
+
 from sklearn.linear_model import LogisticRegressionCV, LinearRegression, RidgeCV
 from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
 from sklearn.preprocessing import PolynomialFeatures
@@ -24,15 +20,6 @@ from scipy.special import expit
 from itertools import combinations
 from sklearn.model_selection import KFold
 
-pandas2ri.activate()
-numpy2ri.activate()
-
-causalweight = rpackages.importr('causalweight')
-mediation = rpackages.importr('mediation')
-Rstats = rpackages.importr('stats')
-base = rpackages.importr('base')
-grf = rpackages.importr('grf')
-plmed = rpackages.importr('plmed')
 
 ALPHAS = np.logspace(-5, 5, 8)
 CV_FOLDS = 5
@@ -81,86 +68,6 @@ def get_interactions(interaction, *args):
         result = np.hstack(variables + (new_vars,))
         return result
 
-
-def plain_IPW(y, t, x, trim=0.01, regularization=True):
-    """
-    plain IPW estimator without mediation
-    """
-    if regularization:
-        cs = ALPHAS
-    else:
-        cs = [np.inf]
-    if len(x.shape) == 1:
-        x = x.reshape(-1, 1)
-    p_x_clf = LogisticRegressionCV(Cs=cs, cv=CV_FOLDS).fit(x, t)
-    p_x = p_x_clf.predict_proba(x)[:, 1]
-    # trimming
-    p_x[p_x < trim] = trim
-    p_x[p_x > 1 - trim] = 1 - trim
-    y1m1 = np.sum(y * t / p_x) / np.sum(t / p_x)
-    y0m0 = np.sum(y * (1 - t) / (1 - p_x)) /\
-        np.sum((1 - t) / (1 - p_x))
-    return y1m1 - y0m0
-
-
-def AIPW(y, t, m, x, clip=0.01, forest=False, crossfit=0, forest_r=False,
-         regularization=True):
-    """
-    AIPW estimator
-    """
-    if regularization:
-        alphas = ALPHAS
-        cs = ALPHAS
-    else:
-        alphas = [0.0]
-        cs = [np.inf]
-    n = len(y)
-    if len(x.shape) == 1:
-        x = x.reshape(-1, 1)
-    if crossfit < 2:
-        train_test_list = [[np.arange(n), np.arange(n)]]
-    else:
-        kf = KFold(n_splits=crossfit)
-        train_test_list = list()
-        for train_index, test_index in kf.split(x):
-            train_test_list.append([train_index, test_index])
-
-    if not forest_r:
-        mu_1x, mu_0x, e_x = [np.zeros(n) for h in range(3)]
-
-        for train_index, test_index in train_test_list:
-            treated_train_index = np.array(
-                list(set(train_index).intersection(np.where(t == 1)[0])))
-            control_train_index = np.array(
-                list(set(train_index).intersection(np.where(t == 0)[0])))
-            if not forest:
-                y_reg_treated = RidgeCV(alphas=alphas, cv=CV_FOLDS)\
-                    .fit(x[treated_train_index, :], y[treated_train_index])
-                y_reg_control = RidgeCV(alphas=alphas, cv=CV_FOLDS)\
-                    .fit(x[control_train_index, :], y[control_train_index])
-                t_prob = LogisticRegressionCV(Cs=cs, cv=CV_FOLDS)\
-                    .fit(x[train_index, :], t[train_index])
-            else:
-                y_reg_treated = RandomForestRegressor(max_depth=3,
-                                                      min_samples_leaf=5)\
-                    .fit(x[treated_train_index, :], y[treated_train_index])
-                y_reg_control = RandomForestRegressor(max_depth=3,
-                                                      min_samples_leaf=5)\
-                    .fit(x[control_train_index, :], y[control_train_index])
-                t_prob = CalibratedClassifierCV(
-                    RandomForestClassifier(max_depth=3, min_samples_leaf=5))\
-                    .fit(x[train_index, :], t[train_index])
-            mu_1x[test_index] = y_reg_treated.predict(x[test_index, :])
-            mu_0x[test_index] = y_reg_control.predict(x[test_index, :])
-            e_x[test_index] = t_prob.predict_proba(x[test_index, :])[:, 1]
-        e_x = np.clip(e_x, clip, 1 - clip)
-        total_effect = np.mean(mu_1x - mu_0x + t * (y - mu_1x) / e_x -
-                               (1 - t) * (y - mu_0x) / (1 - e_x))
-    else:
-        x_r, t_r, y_r = [_convert_array_to_R(uu) for uu in (x, t, y)]
-        cf = grf.causal_forest(x_r, y_r, t_r, num_trees=500)
-        total_effect = grf.average_treatment_effect(cf)[0]
-    return [total_effect] + [None] * 5
 
 
 def huber_IPW(y, t, m, x, w, z, trim, logit, regularization=True, forest=False,
@@ -785,165 +692,3 @@ def multiply_robust_efficient(y, t, m, x, interaction=False,
             indirect_effect_treated,
             np.nan,
             None]
-
-
-def r_mediate(y, t, m, x, interaction=False):
-    """
-    This function calls the R function mediate from the package mediation
-    (https://cran.r-project.org/package=mediation)
-    y       array-like, shape (n_samples)
-            outcome value for each unit, continuous
-
-    t       array-like, shape (n_samples)
-            treatment value for each unit, binary
-
-    m       array-like, shape (n_samples)
-            mediator value for each unit, here m is necessary binary and uni-
-            dimensional
-
-    x       array-like, shape (n_samples, n_features_covariates)
-            covariates (potential confounders) values
-
-    interaction boolean, default=False
-                whether to include interaction terms in the model
-                interactions are terms XT, TM, MX
-    """
-    m = m.ravel()
-    var_names = [[y, 'y'],
-                 [t, 't'],
-                 [m, 'm'],
-                 [x, 'x']]
-    df_list = list()
-    for var, name in var_names:
-        if len(var.shape) > 1:
-            var_dim = var.shape[1]
-            col_names = ['{}_{}'.format(name, i) for i in range(var_dim)]
-            sub_df = pd.DataFrame(var, columns=col_names)
-        else:
-            sub_df = pd.DataFrame(var, columns=[name])
-        df_list.append(sub_df)
-        df = pd.concat(df_list, axis=1)
-    m_features = [c for c in df.columns if ('y' not in c) and ('m' not in c)]
-    y_features = [c for c in df.columns if ('y' not in c)]
-    if not interaction:
-        m_formula = 'm ~ ' + ' + '.join(m_features)
-        y_formula = 'y ~ ' + ' + '.join(y_features)
-    else:
-        m_formula = 'm ~ ' + ' + '.join(m_features +
-                                        [':'.join(p) for p in
-                                         combinations(m_features, 2)])
-        y_formula = 'y ~ ' + ' + '.join(y_features +
-                                        [':'.join(p) for p in
-                                         combinations(y_features, 2)])
-    robjects.globalenv['df'] = df
-    mediator_model = Rstats.lm(m_formula, data=base.as_symbol('df'))
-    outcome_model = Rstats.lm(y_formula, data=base.as_symbol('df'))
-    res = mediation.mediate(mediator_model, outcome_model, treat='t',
-                            mediator='m', boot=True, sims=1)
-
-    relevant_variables = ['tau.coef', 'z1', 'z0', 'd1', 'd0']
-    to_return = [np.array(res.rx2(v))[0] for v in relevant_variables]
-    return to_return + [None]
-
-
-def g_estimator(y, t, m, x):
-    m = m.ravel()
-    var_names = [[y, 'y'],
-                 [t, 't'],
-                 [m, 'm'],
-                 [x, 'x']]
-    df_list = list()
-    for var, name in var_names:
-        if len(var.shape) > 1:
-            var_dim = var.shape[1]
-            col_names = ['{}_{}'.format(name, i) for i in range(var_dim)]
-            sub_df = pd.DataFrame(var, columns=col_names)
-        else:
-            sub_df = pd.DataFrame(var, columns=[name])
-        df_list.append(sub_df)
-        df = pd.concat(df_list, axis=1)
-    m_features = [c for c in df.columns if ('x' in c)]
-    y_features = [c for c in df.columns if ('x' in c)]
-    t_features = [c for c in df.columns if ('x' in c)]
-    m_formula = 'm ~ ' + ' + '.join(m_features)
-    y_formula = 'y ~ ' + ' + '.join(y_features)
-    t_formula = 't ~ ' + ' + '.join(t_features)
-    robjects.globalenv['df'] = df
-    res = plmed.G_estimation(t_formula,
-                             m_formula,
-                             y_formula,
-                             exposure_family='binomial',
-                             data=base.as_symbol('df'))
-    direct_effect = res.rx2('coef')[0]
-    indirect_effect = res.rx2('coef')[1]
-    return [direct_effect + indirect_effect,
-            direct_effect,
-            direct_effect,
-            indirect_effect,
-            indirect_effect,
-            None]
-
-
-def bart(y, t, m, x, tmle=False):
-    x_r, t_r, y_r = [_convert_array_to_R(uu) for uu in (x, t, y)]
-    if tmle:
-        bart_model = brc.bartc(y_r, t_r, x_r, method_rsp='tmle',
-                               estimand='ate')
-    else:
-        bart_model = brc.bartc(y_r, t_r, x_r, method_rsp='p.weight',
-                               estimand='ate')
-    ate = brc.summary_bartcFit(bart_model).rx2('estimates')[0][0]
-    return [ate] + [None] * 5
-
-
-def medDML(y, t, m, x, trim=0.05, order=1):
-    """
-    y       array-like, shape (n_samples)
-            outcome value for each unit, continuous
-
-    t       array-like, shape (n_samples)
-            treatment value for each unit, binary
-
-    m       array-like, shape (n_samples, n_features_mediator)
-            mediator value for each unit, can be continuous or binary, and
-            multi-dimensional
-
-    x       array-like, shape (n_samples, n_features_covariates)
-            covariates (potential confounders) values
-
-    trim    float
-            Trimming rule for discarding observations with extreme
-            conditional treatment or mediator probabilities
-            (or products thereof). Observations with (products of)
-            conditional probabilities that are smaller than trim in any
-            denominator of the potential outcomes are dropped.
-            Default is 0.05.
-
-    order   integer
-            If set to an integer larger than 1, then polynomials of that
-            order and interactions using the power series) rather than the
-            original control variables are used in the estimation of any
-            conditional probability or conditional mean outcome.
-            Polynomials/interactions are created using the Generate.
-            Powers command of the LARF package.
-    """
-    x_r, t_r, m_r, y_r = [base.as_matrix(_convert_array_to_R(uu)) for uu in
-                          (x, t, m, y)]
-    res = causalweight.medDML(y_r, t_r, m_r, x_r, trim=trim, order=order)
-    raw_res_R = np.array(res.rx2('results'))
-    ntrimmed = res.rx2('ntrimmed')[0]
-    return list(raw_res_R[0, :5]) + [ntrimmed]
-
-
-def _convert_array_to_R(x):
-    """
-    converts a numpy array to a R matrix or vector
-    >>> a = np.array([[1, 2, 3], [4, 5, 6]])
-    >>> np.sum(a == np.array(_convert_array_to_R(a)))
-    6
-    """
-    if len(x.shape) == 1:
-        return robjects.FloatVector(x)
-    elif len(x.shape) == 2:
-        return robjects.r.matrix(robjects.FloatVector(x.ravel()),
-                                 nrow=x.shape[0], byrow='TRUE')
